Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Abdul Wahab",email = "wahabbasa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "rich (>=13.7.0,<14.0.0)",
    "instructor (>=1.7.9,<2.0.0)",
    "atomic-agents (>=1.0.26,<2.0.0)",
    "openai (>=1.72.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
